requirements.txt
fpdf2
plotly
altair
streamlit
pandas
gspread
google-auth
google-auth-oauthlib
streamlit-option-menu
python-barcode
pillow
qrcode
python-barcode[images]
